% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{loo_fit}
\alias{loo_fit}
\title{Approximate leave-one-out cross-validation (PSIS-LOO)}
\usage{
loo_fit(cfit)
}
\arguments{
\item{cfit}{A `cure_fit` object.}
}
\description{
Approximate leave-one-out cross-validation (PSIS-LOO)
}
