% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{mcmc_checks}
\alias{mcmc_checks}
\title{MCMC diagnostic summary (R-hat, n_eff, divergences, tree-depth)}
\usage{
mcmc_checks(cfit)
}
\arguments{
\item{cfit}{A `cure_fit` object.}
}
\description{
MCMC diagnostic summary (R-hat, n_eff, divergences, tree-depth)
}
