% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{plot_surv_curve}
\alias{plot_surv_curve}
\title{Population survival curve implied by the cure model}
\usage{
plot_surv_curve(
  cfit,
  t_grid = seq(0.01, 15, by = 0.1),
  main_title = "Population survival (promotion time cure)"
)
}
\arguments{
\item{cfit}{A `cure_fit` object.}

\item{t_grid}{Numeric grid of times at which to evaluate S(t).}

\item{main_title}{Plot title.}
}
\description{
Population survival curve implied by the cure model
}
