% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgetne-package.R
\docType{package}
\name{rgetne-package}
\alias{rgetne}
\alias{rgetne-package}
\title{rgetne: Bayesian Promotion Time Cure Model Toolkit and Interactive App}
\description{
Fits Bayesian Promotion Time Cure Models (Yakovlev & Tsodikov bounded
cumulative hazard) via Stan, supports interactions and per-variable
priors, ships identifiability diagnostics and clinical plots, and
bundles an interactive Shiny app (see [cure()]).
}
\section{Quick start}{

\preformatted{
  library(rgetne)
  cure()                       # launches the Shiny app

  dat  <- prep_data(df_datos_app)
  mod  <- cure_model(dat,
    clonogenic = ~ ki67_percent * periop_therapy + stage_III,
    kinetic    = ~ ki67_percent + periop_therapy)
  cfit <- fit_model(mod)
  summary_coefs(cfit)
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/albertocarm/rgetne}
  \item Report bugs at \url{https://github.com/albertocarm/rgetne/issues}
}

}
\author{
\strong{Maintainer}: Rocio Maintainer \email{rocio@example.com}

}
\keyword{internal}
