% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{stan_model_file}
\alias{stan_model_file}
\title{Path to the bundled Stan model}
\usage{
stan_model_file()
}
\value{
Absolute path to `promotion_time_cure_v2.stan` shipped with the
  package.
}
\description{
Path to the bundled Stan model
}
