% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{summary_coefs}
\alias{summary_coefs}
\title{Coefficient summary on the original (un-scaled) covariate scale}
\usage{
summary_coefs(cfit)
}
\arguments{
\item{cfit}{A `cure_fit` object.}
}
\description{
Coefficient summary on the original (un-scaled) covariate scale
}
