% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{test_censoring}
\alias{test_censoring}
\title{Incremental censoring sensitivity of the cure fraction}
\usage{
test_censoring(cfit, quantiles = c(1, 0.9, 0.8, 0.7), chains = 2, iter = 2000)
}
\arguments{
\item{cfit}{A `cure_fit` object.}

\item{quantiles}{Follow-up quantiles to refit at.}

\item{chains, iter}{MCMC settings used for the refits.}
}
\description{
Incremental censoring sensitivity of the cure fraction
}
