% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{test_correlation}
\alias{test_correlation}
\title{Posterior cross-correlation between clonogenic and kinetic coefficients}
\usage{
test_correlation(cfit)
}
\arguments{
\item{cfit}{A `cure_fit` object.}
}
\description{
Posterior cross-correlation between clonogenic and kinetic coefficients
}
